Removed volumes from base Resource & added to Service only
Moved volumes config/spec/transform to the common modules in anticipation of future use
