+++ src/dependency-manager/src/spec/transform/service-transform.ts
@@ -2,7 +2,7 @@ import { ComponentInstanceMetadata } from '../../config/component-config';
 import { ServiceConfig, ServiceInterfaceConfig } from '../../config/service-config';
 import { transformDictionary } from '../../utils/dictionary';
 import { ServiceSpec, ServiceInterfaceSpec } from '../service-spec';
-import { transformLivenessProbeSpec } from './common-transform';
+import { transformLivenessProbeSpec, transformVolumeSpec } from './common-transform';
 import { transformResourceSpec } from './resource-transform';
 import { transformSidecarSpec } from './sidecar-transform';
 
@@ -23,6 +23,7 @@ export const transformServiceSpec = (key: string, spec: ServiceSpec, component_r
     interfaces: transformDictionary(transformInterfaceSpec, spec.interfaces),
     sidecars: transformDictionary(transformSidecarSpec, spec.sidecars, component_ref, tag, instance_metadata),
     liveness_probe: transformLivenessProbeSpec(spec.liveness_probe, resource_config.environment),
+    volumes: transformDictionary(transformVolumeSpec, spec.volumes),
     replicas: spec.replicas || 1,
     scaling: spec.scaling,
   };
